Adding support for variable allocated textures to the GL 4.1 backend
